#8.10 - #8.15: relationship

+++ src/controllers/videoController.js
@@ -1,3 +1,4 @@
+import User from "../models/User";
 import Video from "../models/Video";
 
 export const home = async (req, res) => {
@@ -13,7 +14,7 @@ export const home = async (req, res) => {
 
 export const watch = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner");
 
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found" });
@@ -23,11 +24,20 @@ export const watch = async (req, res) => {
 };
 
 export const getEdit = async (req, res) => {
-  const { id } = req.params;
+  const {
+    params: { id },
+    session: {
+      user: { _id },
+    },
+  } = req;
   const video = await Video.findById(id);
 
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found" });
+  }
+
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
 
   return res.render("edit", { pageTitle: `Edit ${video.title}`, video });
@@ -56,16 +66,26 @@ export const getUpload = (req, res) => {
 };
 
 export const postUpload = async (req, res) => {
-  const { path: fileUrl } = req.file;
-  const { title, description, hashtags } = req.body;
+  const {
+    session: {
+      user: { _id },
+    },
+    file: { path: fileUrl },
+    body: { title, description, hashtags },
+  } = req;
 
   try {
-    await Video.create({
+    const video = await Video.create({
       title,
       description,
       fileUrl,
+      owner: _id,
       hashtags: Video.formatHashtags(hashtags),
     });
+
+    const user = await User.findById(_id);
+    user.videos.push(video.id);
+    user.save();
 
     return res.redirect("/");
   } catch (error) {
@@ -77,8 +97,16 @@ export const postUpload = async (req, res) => {
   }
 };
 
-export const deleteVideo = async (rsq, res) => {
-  const { id } = req.params;
+export const deleteVideo = async (req, res) => {
+  const {
+    params: { id },
+    session: { _id },
+  } = req;
+
+  const video = await Video.findById(id);
+
+  if (video && String(video.owner) !== String(_id))
+    return res.status(403).redirect("/");
 
   await Video.findByIdAndDelete(id);
